Trait oak_attestation_types::endorser::Endorser

pub trait Endorser: Send + Sync {
    // Required method
    fn endorse(&self, evidence: Option<&Evidence>) -> Result<Endorsements>;
}

Trait that provides the ability to read attestation endorsements.

https://datatracker.ietf.org/doc/html/rfc9334#name-endorser-reference-value-pr

Required Methods

fn endorse(&self, evidence: Option<&Evidence>) -> Result<Endorsements>

Generate an endorsement.

Evidence argument is optional since it may be required for endorsement generation in some use-cases.