Trait oak_attestation_types::util::Serializable

pub trait Serializable: Sized {
    // Required methods
    fn deserialize(bytes: &[u8]) -> Result<Self>;
    fn serialize(self) -> Vec<u8>;
}

Trait for passing incomplete evidence between layers of software components.

For example, in DICE it is used to pass the DiceData containing the certificate authority private key, and for TDX it is used to pass an unfinished EventLog.

Required Methods

fn deserialize(bytes: &[u8]) -> Result<Self>

fn serialize(self) -> Vec<u8>

Object Safety

This trait is not object safe.